# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "get": {
        "description": "The storefront listing. Public, but privilege-aware: **sending a bearer token that carries `products.read` changes what comes back**, since only that caller may use the `status` filter and therefore see DRAFT or ARCHIVED products. Without a token, or with one lacking the permission, the listing is ACTIVE products only.\n\nOpenAPI cannot express \"optional bearer\", so this route is documented as unauthenticated — but a token is accepted, and a token that is offered and rejected is a 401 rather than a silent downgrade to anonymous.",
        "operationId": "ProductsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "description": "Values above 100 are clamped, not rejected.",
            "schema": {
              "minimum": 1,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Category slug.",
            "schema": {
              "example": "camisetas",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive partial match on the name. Not a search engine.",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "REQUIRES the `products.read` permission — without it this is a 403, not a silently ignored filter. `all` means every status.",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "ACTIVE",
                "ARCHIVED",
                "all"
              ]
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Ordering, applied before pagination. Omit for `newest` (createdAt desc), the historical default. Every ordering breaks ties on id so pages never repeat or drop an item.",
            "schema": {
              "default": "newest",
              "type": "string",
              "enum": [
                "newest",
                "price_asc",
                "price_desc",
                "name_asc"
              ]
            }
          },
          {
            "name": "minPriceCents",
            "required": false,
            "in": "query",
            "description": "Inclusive lower bound, integer cents. Greater than `maxPriceCents` is a 400, not an empty list.",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "maxPriceCents",
            "required": false,
            "in": "query",
            "description": "Inclusive upper bound, integer cents.",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedProductsResponse"
                }
              }
            }
          },
          "400": {
            "description": "A query parameter failed validation — a non-integer `page`, or a `status` outside the enum.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The `status` filter was used without the `products.read` permission. Asking a privileged question is refused rather than silently answered as if ACTIVE had been requested.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "List products",
        "tags": [
          "products"
        ]
      },
      "post": {
        "description": "Products are born DRAFT and stay off the storefront until moved to ACTIVE. Omitting `slug` generates one from the name, adding a numeric suffix on collision; sending one that is already taken is a 409 instead, because a caller who chose the slug wants that slug.\n\nSend `variants` in display order — `[{ \"label\": \"P\" }, { \"label\": \"M\" }, …]` — and their `position` follows the array. Omit it and the product is born with exactly one variant labelled `Único`: a product is never created without at least one, because stock lives on the variant and a product with none would be unbuyable.",
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "`priceCents` is not an integer above zero, a category id is not a UUID, an image is not a URL, the slug is not lowercase-hyphenated, or two variants share a label.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No bearer token, or the token is expired or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the `products.create` permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "One of the `categoryIds` does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The slug sent explicitly is already taken.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a product",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{idOrSlug}": {
      "get": {
        "description": "Accepts either the id or the slug in the same path segment — ids are UUIDs and slugs never look like one, so the two cannot collide. Prefer the id for links that must survive a rename.\n\nA non-ACTIVE product is a 404 to anyone without `products.read`, identical to a product that does not exist. A 403 would confirm to someone probing slugs that an unreleased product is sitting there.",
        "operationId": "ProductsController_get",
        "parameters": [
          {
            "name": "idOrSlug",
            "required": true,
            "in": "path",
            "description": "The product UUID or its slug.",
            "schema": {
              "example": "camiseta-azul",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such product — or it is DRAFT/ARCHIVED and the caller lacks `products.read`. The two are deliberately indistinguishable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Get one product",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}/variants": {
      "post": {
        "description": "Adds one variant. `position` defaults to the end of the list; `stockQuantity` defaults to 0, which is a real state — the size exists and has none left.\n\nThere is deliberately no route to rename, reorder or remove a variant. Adding cannot invalidate anything; removing has to decide what happens to a size somebody already bought, and that is a policy decision rather than a detail (docs/specs/product-variants.md).",
        "operationId": "ProductsController_addVariant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVariantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The whole product, with its variants in display order.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "`label` is empty or too long, or `position` is negative.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No bearer token, or the token is expired or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the `products.update` permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No product with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "This product already has a variant with that label.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a size to a product",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}/variants/{variantId}/stock": {
      "patch": {
        "description": "An inventory correction — \"the shelf holds N of this size\" — not a delta. Selling is the other path: checkout decrements the variant atomically inside its transaction, and nothing here is involved in that.\n\nStock belongs to the variant, never to the product: `ProductResponse.stockQuantity` is the sum across sizes, computed on read, so it cannot drift from what checkout takes.\n\nBe aware this is last-write-wins against a concurrent sale, which is accepted for v1 (docs/specs/catalog.md).",
        "operationId": "ProductsController_setStock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetStockDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "`quantity` is negative or not an integer.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No bearer token, or the token is expired or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the `products.update` permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such product, or that variant does not belong to it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set one size's stock to an absolute quantity",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}": {
      "patch": {
        "description": "Every field is optional; omitted fields are left alone. `categoryIds` is the exception to \"partial\": absent leaves the associations untouched, present **replaces the whole set**.\n\nThis is also how a product reaches ACTIVE — there is no separate publish route.",
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "A field failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No bearer token, or the token is expired or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the `products.update` permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No product with that id, or a category id does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The new slug is already taken.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a product",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "description": "**Archives rather than deletes.** Orders reference products forever, so removing the row would break financial history; the product moves to ARCHIVED, leaves the storefront and refuses new sales, and the row stays.\n\nThe response is the archived product, not an empty body.",
        "operationId": "ProductsController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "401": {
            "description": "No bearer token, or the token is expired or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the `products.delete` permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No product with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archive a product",
        "tags": [
          "products"
        ]
      }
    },
...
  },
  "info": {
    "title": "commerce-core",
...
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "The `accessToken` from /auth/login, /auth/refresh or the Google callback."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 409,
            "description": "Mirrors the HTTP status code."
          },
          "message": {
            "description": "A human-readable reason, or one entry per failed validation rule.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "examples": [
              "Order is PAID; only a CREATED order can be paid",
              [
                "priceCents must be an integer number",
                "name should not be empty"
              ]
            ]
          },
          "error": {
            "type": "string",
            "example": "Conflict",
            "description": "The status code's canonical name."
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "PaginatedProductsResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductResponse"
            }
          },
          "total": {
            "type": "number",
            "description": "Total matching the filters, not the page size.",
            "example": 137
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "perPage": {
            "type": "number",
            "description": "Clamped to 100.",
            "example": 20
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "perPage"
        ]
      },
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "example": "Camiseta Azul"
          },
          "slug": {
            "type": "string",
            "maxLength": 200,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "description": "Omit to generate one from the name, with a numeric suffix on collision. Send one that is taken and the answer is 409 — a caller who picked the slug wants that slug.",
            "example": "camiseta-azul"
          },
          "description": {
            "type": "string",
            "maxLength": 5000
          },
          "priceCents": {
            "type": "number",
            "minimum": 1,
            "description": "Integer cents, greater than zero. Never a float.",
            "example": 7990
          },
          "imageUrls": {
            "maxItems": 20,
            "description": "Plain URLs — this API does not host images in v1.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "ARCHIVED"
            ],
            "default": "DRAFT",
            "description": "Products are born DRAFT and stay off the storefront until ACTIVE."
          },
          "variants": {
            "maxItems": 50,
            "description": "The sellable sizes, in display order. Omit and the product gets one variant labelled `Único` at zero stock — a product is never created without at least one. Labels must be unique within the product.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateVariantDto"
            }
          },
          "weightGrams": {
            "type": "number",
            "minimum": 1,
            "description": "Unit weight in grams, used to price freight. Omit and quotes fall back to a configured default — which the store pays for when the guess is low, so it is worth filling in.",
            "example": 180
          },
          "categoryIds": {
            "maxItems": 50,
            "description": "Category ids. A product in no category is valid.",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "name",
          "priceCents"
        ]
      },
      "SetStockDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "minimum": 0,
            "description": "The absolute count on the shelf, not a delta. Selling is the other path — checkout decrements atomically on its own.",
            "example": 12
          }
        },
        "required": [
          "quantity"
        ]
      },
...
    }
  }
}